removed unused crates, improved examples/echo.rs, added a some event fields
