+++ Cargo.toml
@@ -19,6 +19,4 @@ serde = "0.9"
 serde_derive = "0.9"
 serde_json = "0.9"
 tokio-core = "0.1"
-
-[dev-dependencies]
-json = "0.11.8"
+rpassword = "0.4.2"
